+++ frontend/components/layout/StagingBanner.stories.js
@@ -1,0 +1,11 @@
+import { storiesOf } from '@storybook/vue';
+import StagingBanner from '~/components/layout/StagingBanner';
+
+const common = {
+    components: { StagingBanner },
+};
+
+storiesOf('Layout/StagingBanner', module).add('Default', () => ({
+    ...common,
+    template: `<StagingBanner />`,
+}));
