--- frontend/components/layout/StagingHeader.stories.js
@@ -1,11 +1,0 @@
-import { storiesOf } from '@storybook/vue';
-import StagingHeader from '~/components/layout/StagingHeader';
-
-const common = {
-    components: { StagingHeader },
-};
-
-storiesOf('Layout/StagingHeader', module).add('Default', () => ({
-    ...common,
-    template: `<StagingHeader />`,
-}));
